+++ tencentcloud/as/v20180419/models.go
@@ -3529,7 +3529,7 @@ type LifecycleCommand struct {
 	// 如果未提供该参数取值，将使用 Command 的 DefaultParameters 进行替换。
 	// 自定义参数最多20个。自定义参数名称需符合以下规范：字符数目上限64，可选范围【a-zA-Z0-9-_】。
 	// 注意：此字段可能返回 null，表示取不到有效值。
-	Parameters []*string `json:"Parameters,omitempty" name:"Parameters"`
+	Parameters *string `json:"Parameters,omitempty" name:"Parameters"`
 }
 
 type LifecycleHook struct {