+++ tencentcloud/clb/v20180317/models.go
@@ -2100,10 +2100,10 @@ type DeleteRuleRequestParams struct {
 	// 要删除的转发规则的ID组成的数组。
 	LocationIds []*string `json:"LocationIds,omitempty" name:"LocationIds"`
 
-	// 要删除的转发规则的域名，如果是多域名，可以指定多域名列表中的任意一个。已提供LocationIds参数时本参数不生效。
+	// 要删除的转发规则的域名，如果是多域名，可以指定多域名列表中的任意一个。
 	Domain *string `json:"Domain,omitempty" name:"Domain"`
 
-	// 要删除的转发规则的转发路径，已提供LocationIds参数时本参数不生效。
+	// 要删除的转发规则的转发路径。
 	Url *string `json:"Url,omitempty" name:"Url"`
 
 	// 监听器下必须配置一个默认域名，当需要删除默认域名时，可以指定另一个域名作为新的默认域名，如果新的默认域名是多域名，可以指定多域名列表中的任意一个。
@@ -2122,10 +2122,10 @@ type DeleteRuleRequest struct {
 	// 要删除的转发规则的ID组成的数组。
 	LocationIds []*string `json:"LocationIds,omitempty" name:"LocationIds"`
 
-	// 要删除的转发规则的域名，如果是多域名，可以指定多域名列表中的任意一个。已提供LocationIds参数时本参数不生效。
+	// 要删除的转发规则的域名，如果是多域名，可以指定多域名列表中的任意一个。
 	Domain *string `json:"Domain,omitempty" name:"Domain"`
 
-	// 要删除的转发规则的转发路径，已提供LocationIds参数时本参数不生效。
+	// 要删除的转发规则的转发路径。
 	Url *string `json:"Url,omitempty" name:"Url"`
 
 	// 监听器下必须配置一个默认域名，当需要删除默认域名时，可以指定另一个域名作为新的默认域名，如果新的默认域名是多域名，可以指定多域名列表中的任意一个。